Break `sqlrepo` into files by repo interface
